+++ bot.sh
@@ -1,0 +1,6 @@
+function bot
+{
+    node --no-warnings main.js
+    bot
+}
+bot
