+++ commands/music/old.i/yamusic2.js
