+++ commands/music/old.i/yandexmusic.js
