+++ package.json
@@ -17,6 +17,7 @@
     "libsodium-wrappers": "^0.7.11",
     "net": "^1.0.2",
     "openai": "^3.2.1",
+    "pawan_openai": "^1.1.2",
     "play-dl": "^1.9.6",
     "utf-8-validate": "^6.0.3",
     "youtube-sr": "^4.3.4"
